feat: Implement document readers for PDF and DOCX formats
- Added `read_pdf_text` function in `backend/app/features/adapters/pdf/reader.py` to extract text from PDF files using `pypdf`.
- Introduced `read_docx_text` function in `backend/app/features/adapters/docx/reader.py` to extract text from DOCX files using multiple libraries (`python-docx`, `docx2txt`, `docx2python`) for better coverage.
- Updated `ResumeService` to utilize the new document readers for text extraction.
- Removed deprecated text extraction functions from `backend/app/features/resumes/text_extractors.py` and redirected to the new reader functions.
- Added parsing utilities in `backend/app/features/parsing` for better organization and modularity, including sections for normalizing text, parsing contact information, skills, and education entries.
- Enhanced `CVParser` to delegate parsing logic to the new `ParserCore` class for improved maintainability.
- Updated test configuration in `conftest.py` to ensure proper path resolution for tests.
