+++ backend/app/features/parsing/__init__.py
@@ -1,0 +1,10 @@
+"""Parsing utilities for resume extraction.
+
+Modules:
+- normalizers: text cleanup and unicode normalization
+- sections: split text into resume-like sections
+- contact: email/phone patterns and extractors
+- skills: curated skills detection logic
+- education: education entries parsing
+- parser_core: orchestrates parsing into ResumeParsedJSON
+"""
