+++ backend/app/features/resumes/service.py
@@ -6,7 +6,8 @@ from pathlib import Path
 from fastapi import UploadFile, HTTPException, status
 from app.core.config import MAX_UPLOAD_SIZE, UPLOAD_DIR
 from app.utils.sanitize import safe_filename
-from .text_extractors import extract_text_from_pdf, extract_text_from_docx
+from app.features.adapters.pdf.reader import read_pdf_text
+from app.features.adapters.docx.reader import read_docx_text
 from .cv_parser import CVParser
 from .jsonb_models import ResumeParsedJSON
 from .security import verify_magic_bytes, SUPPORTED_MIME
@@ -79,6 +80,6 @@ class ResumeService:
 
     def _extract_text(self, path: Path, content_type: str) -> str:
         if content_type == "application/pdf":
-            return extract_text_from_pdf(path)
+            return read_pdf_text(path)
         # else DOCX
-        return extract_text_from_docx(path)
+        return read_docx_text(path)
